refactor: Organize tables
