+++ src/app/_components/PortalComponents/Portal/recruitment/columns.tsx
@@ -6,7 +6,7 @@ import { createColumnHelper } from "@tanstack/react-table";
 
 import EditRecruitmentFormCell from "../../EditRecruitmentFormCell";
 import DeleteForm from "../../EditRecruitmentFormCell/DeleteForm";
-import { RecruitmentForm } from "../RecruitmentTable";
+import { RecruitmentForm } from "./RecruitmentTable";
 
 const columnHelper = createColumnHelper<RecruitmentForm>();
 export const columns = [
