+++ src/app/_components/PortalComponents/Portal/sponsors/SponsorsTable.tsx
@@ -2,8 +2,8 @@ import { useUser } from "@/app/_hooks/useUser";
 import { type RouterOutputs } from "@/trpc/react";
 import { type ColumnDef, type VisibilityState } from "@tanstack/react-table";
 
-import EntityTable from "./EntityTable";
-import { columns } from "./sponsors/columns";
+import EntityTable from "../EntityTable";
+import { columns } from "./columns";
 
 export type Sponsor = RouterOutputs["portal"]["getSponsorsList"][number];
